requirement.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dict.sorted==1.0.0
Django==3.2.16
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==5.2.1
drf-toolbox==0.1.7
drf-yasg==1.21.4
environ==1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
psycopg2==2.9.4
psycopg2-binary==2.9.5
PyJWT==2.5.0
pyparsing==3.0.9
pytz==2022.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
